+++ custom_model_runner/datarobot_drum/drum/gpu_predictors/base.py
@@ -27,7 +27,7 @@ from datarobot_drum import RuntimeParameters
 from datarobot_drum.drum.adapters.model_adapters.python_model_adapter import PythonModelAdapter
 from datarobot_drum.drum.adapters.model_adapters.python_model_adapter import RawPredictResponse
 from datarobot_drum.drum.common import SupportedPayloadFormats
-from datarobot_drum.drum.enum import CUSTOM_FILE_NAME, EnvVarNames
+from datarobot_drum.drum.enum import CUSTOM_FILE_NAME, EnvVarNames, TargetType
 from datarobot_drum.drum.enum import LOGGER_NAME_PREFIX
 from datarobot_drum.drum.enum import CustomHooks
 from datarobot_drum.drum.enum import PayloadFormat
@@ -74,7 +74,9 @@ class BaseOpenAiGpuPredictor(BaseLanguagePredictor):
         # chat input fields
         self.system_prompt_value = self.get_optional_parameter("system_prompt")
         self.user_prompt_column = self.get_optional_parameter("prompt_column_name", "promptText")
-        self.served_model_name = self.get_optional_parameter("served_model_name", self.DEFAULT_MODEL_NAME)
+        self.served_model_name: str = self.get_optional_parameter(
+            "served_model_name", self.DEFAULT_MODEL_NAME
+        )
 
         # completions configuration can be changed with Runtime parameters
         self.max_tokens = int(self.get_optional_parameter("max_tokens", 0)) or None
@@ -82,7 +84,7 @@ class BaseOpenAiGpuPredictor(BaseLanguagePredictor):
         self.temperature = self.get_optional_parameter("temperature") or None
 
         # used to load custom model hooks
-        self.python_model_adapter = None
+        self.python_model_adapter: PythonModelAdapter = None
         # report deployment status events to DataRobot
         self.verify_ssl = self.get_optional_parameter("verifySSL", True)
         self.status_reporter: MLOpsStatusReporter = None
@@ -109,7 +111,7 @@ class BaseOpenAiGpuPredictor(BaseLanguagePredictor):
     def _chat(self, completion_create_params, association_id):
         # Force the incoming model name to to match the expected model name because the
         # name isn't very applicable to BYO LLMs.
-        model_name = completion_create_params.pop("model", self.served_model_name)
+        model_name = completion_create_params.pop("model") or self.served_model_name
         completion_create_params["model"] = model_name
         return self.ai_client.chat.completions.create(**completion_create_params)
 
@@ -251,6 +253,13 @@ class BaseOpenAiGpuPredictor(BaseLanguagePredictor):
         return default_value
 
     def _predict(self, **kwargs) -> RawPredictResponse:
+        if self.target_type != TargetType.TEXT_GENERATION:
+            raise DrumCommonException(
+                "The default predict() implementation for this Predictor requires a Text "
+                "Generation target type. To support other target types, you must implement "
+                "a custom hook."
+            )
+
         data = kwargs.get(StructuredDtoKeys.BINARY_DATA)
         if isinstance(data, bytes):
             data = data.decode("utf8")